pyproject.toml
[project]
name = "linkedin-post-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.84.0",
    "beautifulsoup4>=4.14.3",
    "instructor>=1.14.5",
    "numpy>=2.4.2",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "selenium>=4.41.0",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
